Adjustments from First Pull Request
