+++ app/src/main/AndroidManifest.xml
@@ -9,7 +9,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".main.view.activity.MainActivity"
+        <activity android:name=".actualmonth.view.activity.ActualMonthActivity"
             android:label="MAIN"
             android:theme="@style/AppTheme.NoActionBar">
             <intent-filter>
