+++ app/src/main/java/org/upv/ccupeiro/contadroid/actualmonth/view/fragments/TabExpensesNotPaidFragment.java
@@ -1,4 +1,4 @@
-package org.upv.ccupeiro.contadroid.main.view.fragments;
+package org.upv.ccupeiro.contadroid.actualmonth.view.fragments;
 
 import android.os.Bundle;
 import android.support.annotation.Nullable;
@@ -13,8 +13,8 @@ import android.view.ViewGroup;
 import org.upv.ccupeiro.contadroid.R;
 import org.upv.ccupeiro.contadroid.common.model.Expense;
 import org.upv.ccupeiro.contadroid.common.model.ExpensesGroup;
-import org.upv.ccupeiro.contadroid.main.model.CardExpense;
-import org.upv.ccupeiro.contadroid.main.view.adapter.MainCardAdapter;
+import org.upv.ccupeiro.contadroid.actualmonth.model.CardExpense;
+import org.upv.ccupeiro.contadroid.actualmonth.view.adapter.MainCardAdapter;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -22,20 +22,21 @@ import java.util.List;
 import butterknife.BindView;
 import butterknife.ButterKnife;
 
-public class TabExpensesPaidFragment extends Fragment {
+public class TabExpensesNotPaidFragment extends Fragment {
     @Nullable
-    @BindView(R.id.rv_card_paid_expenses)
-    RecyclerView rvPaidExpenses;
+    @BindView(R.id.rv_card_not_paid_expenses)
+    RecyclerView rvNotPaidExpenses;
 
     MainCardAdapter adapter;
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        View rootView = inflater.inflate(R.layout.fragment_tab_expenses_paid, container, false);
+        View rootView = inflater.inflate(R.layout.fragment_tab_expenses_not_paid, container, false);
         initializaButterKnife(rootView);
         prepareAdapter();
         return rootView;
+
     }
 
     private void initializaButterKnife(View view){
@@ -44,19 +45,15 @@ public class TabExpensesPaidFragment extends Fragment {
 
     private void prepareAdapter() {
         adapter = new MainCardAdapter(mockListCardsPaid());
-        rvPaidExpenses.setLayoutManager(new LinearLayoutManager(getContext()));
-        rvPaidExpenses.setAdapter(adapter);
+        rvNotPaidExpenses.setLayoutManager(new LinearLayoutManager(getContext()));
+        rvNotPaidExpenses.setAdapter(adapter);
     }
-
-
 
     private List<CardExpense> mockListCardsPaid(){
         List<CardExpense> cardList = new ArrayList<>();
         List<Expense> expenseList = new ArrayList<>();
-        expenseList.add(new Expense("Nombre",20.0f, ExpensesGroup.INCOME));
-        expenseList.add(new Expense("Nombre","Descripción 1",20.0f, ExpensesGroup.INCOME));
-        expenseList.get(0).setPaid(true);
-        expenseList.get(1).setPaid(true);
+        expenseList.add(new Expense("Nombre sin Pago",150.0f, ExpensesGroup.INCOME));
+        expenseList.add(new Expense("Nombre sin Pago","Descripción sin pagar",10.0f, ExpensesGroup.INCOME));
         cardList.add(new CardExpense(
                 ResourcesCompat.getDrawable(getResources(),R.drawable.icon_income,null),
                 getResources().getString(R.string.group_income),
@@ -65,21 +62,6 @@ public class TabExpensesPaidFragment extends Fragment {
         cardList.add(new CardExpense(
                 ResourcesCompat.getDrawable(getResources(),R.drawable.icon_home,null),
                 getResources().getString(R.string.group_home),
-                expenseList
-        ));
-        cardList.add(new CardExpense(
-                ResourcesCompat.getDrawable(getResources(),R.drawable.icon_transport,null),
-                getResources().getString(R.string.group_transport),
-                expenseList
-        ));
-        cardList.add(new CardExpense(
-                ResourcesCompat.getDrawable(getResources(),R.drawable.icon_food,null),
-                getResources().getString(R.string.group_food),
-                expenseList
-        ));
-        cardList.add(new CardExpense(
-                ResourcesCompat.getDrawable(getResources(),R.drawable.icon_shoppings,null),
-                getResources().getString(R.string.group_shopping),
                 expenseList
         ));
         cardList.add(new CardExpense(
@@ -94,7 +76,5 @@ public class TabExpensesPaidFragment extends Fragment {
         ));
         return cardList;
     }
-
-
 
 }
