+++ app/src/main/java/org/upv/ccupeiro/contadroid/common/model/ExpensesGroup.java
@@ -1,11 +1,5 @@
 package org.upv.ccupeiro.contadroid.common.model;
 
-import org.upv.ccupeiro.contadroid.R;
-
-/**
- * Created by Carlos on 01/08/2017.
- */
-
 public enum ExpensesGroup {
     INCOME,
     HOME,
