+++ app/src/main/java/org/upv/ccupeiro/contadroid/common/utils/Constant.java
@@ -1,0 +1,5 @@
+package org.upv.ccupeiro.contadroid.common.utils;
+
+public class Constant {
+    public static final String FORMAT_MONEY = "%.2f €";
+}
