--- app/src/main/res/drawable/circle.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android" android:shape="oval" >
-    <size android:width="100dp"
-        android:height="100dp"/>
-    <stroke
-        android:width="3dp"
-        android:color="@android:color/white"/>
-    <solid android:color="@android:color/white"/>
-</shape>
