+++ app/src/main/res/drawable/icon_transport.xml
@@ -2,18 +2,18 @@
 <layer-list xmlns:android="http://schemas.android.com/apk/res/android">
     <item>
         <shape xmlns:android="http://schemas.android.com/apk/res/android" android:shape="oval" >
-            <size android:width="24dp"
-                android:height="24dp"/>
+            <size android:width="@dimen/circle_icon_size"
+                android:height="@dimen/circle_icon_size"/>
             <stroke
-                android:width="3dp"
-                android:color="@color/icon_background_transport_color"/>
-            <solid android:color="@color/icon_background_transport_color"/>
+                android:width="@dimen/circle_icon_stroke_width"
+                android:color="@color/iconBackgroundTransportColor"/>
+            <solid android:color="@color/iconBackgroundTransportColor"/>
         </shape>
     </item>
     <item android:drawable="@drawable/bus"
-        android:bottom="2dp"
-        android:top="2dp"
-        android:right="2dp"
-        android:left="2dp"
+        android:bottom="@dimen/icon_drawable_margin"
+        android:top="@dimen/icon_drawable_margin"
+        android:right="@dimen/icon_drawable_margin"
+        android:left="@dimen/icon_drawable_margin"
         android:gravity="center"/>
 </layer-list>
