+++ app/src/main/res/drawable/plus.xml
@@ -1,8 +1,8 @@
 <!-- drawable/plus.xml -->
 <vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:height="24dp"
-    android:width="24dp"
-    android:viewportWidth="24"
-    android:viewportHeight="24">
-    <path android:fillColor="#fff" android:pathData="M19,13H13V19H11V13H5V11H11V5H13V11H19V13Z" />
+    android:height="@dimen/icon_size"
+    android:width="@dimen/icon_size"
+    android:viewportWidth="@dimen/icon_viewport_size"
+    android:viewportHeight="@dimen/icon_viewport_size">
+    <path android:fillColor="@color/iconColor" android:pathData="M19,13H13V19H11V13H5V11H11V5H13V11H19V13Z" />
 </vector>
