--- app/src/main/res/drawable/side_nav_bar.xml
@@ -1,9 +1,0 @@
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="rectangle">
-    <gradient
-        android:angle="135"
-        android:centerColor="#54cac8"
-        android:endColor="#1D9090"
-        android:startColor="#7acecc"
-        android:type="linear" />
-</shape>
