+++ app/src/main/res/layout/fragment_tab_expenses_not_paid.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="org.upv.ccupeiro.contadroid.main.view.fragments.TabExpensesNotPaidFragment">
+    tools:context="org.upv.ccupeiro.contadroid.actualmonth.view.fragments.TabExpensesNotPaidFragment">
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/rv_card_not_paid_expenses"
