+++ app/src/main/res/layout/fragment_tab_expenses_paid.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="org.upv.ccupeiro.contadroid.main.view.fragments.TabExpensesPaidFragment">
+    tools:context="org.upv.ccupeiro.contadroid.actualmonth.view.fragments.TabExpensesPaidFragment">
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/rv_card_paid_expenses"
