+++ app/src/main/res/values/strings.xml
@@ -11,9 +11,9 @@
     <string name="navigation_drawer_menu_login">Iniciar Sesion</string>
 
 
-    <string name="main.title">Mes Actual</string>
-    <string name="main.tab.expenses_paid">Realizados</string>
-    <string name="main.tab.expenses_not_paid">Pendientes</string>
+    <string name="main_title">Mes Actual</string>
+    <string name="main_tab_expenses_paid">Realizados</string>
+    <string name="main_tab_expenses_not_paid">Pendientes</string>
 
 
     <string name="action_settings">Settings</string>
@@ -25,9 +25,5 @@
     <string name="group_shopping">Compras</string>
     <string name="group_leisure">Ocio</string>
     <string name="group_other">Otros</string>
-
-    <!-- TODO: Remove or change this placeholder text -->
-    <string name="hello_blank_fragment">Hello blank fragment</string>
-
 
 </resources>
